1.6.0

+++ greenbtc/consensus/pos_quality.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from greenbtc.util.ints import uint64
 
 # The actual space in bytes of a plot, is _expected_plot_size(k) * UI_ACTUAL_SPACE_CONSTANT_FACTO
@@ -16,4 +14,4 @@ def _expected_plot_size(k: int) -> uint64:
     is necessary to store the entries in the plot.
     """
 
-    return uint64(((2 * k) + 1) * (2 ** (k - 1)))
+    return ((2 * k) + 1) * (2 ** (k - 1))
